Apply lint

+++ caw/src/sample.rs
@@ -15,7 +15,7 @@ fn parse_wav(buffer: &[u8]) -> Vec<f64> {
         .map(|chunk| {
             let channel_mean = chunk.iter().map(|&x| x as i64).sum::<i64>()
                 / chunk.len() as i64;
-            (channel_mean as f64 / max_value as f64) as f64
+            channel_mean as f64 / max_value as f64
         })
         .collect::<Vec<_>>();
     data_f64
